Separate out libauth.h from auth.h
