+++ src/lib/Libauth/gss/pbs_gss.c
@@ -50,8 +50,7 @@
 #include <gssapi.h>
 #include <krb5.h>
 #include "pbs_ifl.h"
-#include "log.h"
-#include "auth.h"
+#include "libauth.h"
 
 #if defined(KRB5_HEIMDAL)
 #define PBS_GSS_MECH_OID GSS_KRB5_MECHANISM
